feat: Renamed AI Module to AI Corner Module
BREAKING CHANGE:  Renamed AI Module to AI Corner Module
